Update dependencies to 19w11b versions and change package
