+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).  
 
+## [0.5.2] - 2019-03-14
+
+### Changed
+ - Dependencies to Minecraft 19w11b versions
+ - Mod package root to `me.branchpanic.mods` to reflect new domain
+
 ## [0.5.1] - 2019-03-01
 
 ### Fixed
