+++ README.md
@@ -1,6 +1,6 @@
 # stockpile
 
-[![Build Status](https://travis-ci.org/notjoe7F/stockpile.svg?branch=master)](https://travis-ci.org/notjoe7F/stockpile)
+[![Build Status](https://travis-ci.org/branchpanic/stockpile.svg?branch=master)](https://travis-ci.org/branchpanic/stockpile)
 [![CurseForge](http://cf.way2muchnoise.eu/versions/stockpile_latest.svg?style=flat)](https://minecraft.curseforge.com/projects/stockpile)
 
 > A Minecraft mod all about storage.
