+++ gradle.properties
@@ -1,13 +1,13 @@
 org.gradle.jvmargs=-Xmx1G
 
 # Mod version
-stockpile_version=0.5.1
+stockpile_version=0.5.2
 
 # Fabric toolchain versions
-minecraft_version=19w09a
-yarn_mappings=19w09a.1
+minecraft_version=19w11b
+yarn_mappings=19w11b.3
 loader_version=0.3.7.109
 
 # Mod dependency versions
-fabric_version=0.2.3.107
+fabric_version=0.2.3.111
 scala_adapter_version=0.1.0
