+++ src/main/scala/me/branchpanic/mods/stockpile/block/StockpileProperties.scala
@@ -1,0 +1,7 @@
+package me.branchpanic.mods.stockpile.block
+
+import net.minecraft.state.property.BooleanProperty
+
+object StockpileProperties {
+  final val IsOpen = BooleanProperty.create("is_open")
+}
