--- src/main/scala/notjoe/stockpile/StockpileClientInitializer.scala
@@ -1,14 +1,0 @@
-package notjoe.stockpile
-
-import net.fabricmc.api.{ClientModInitializer, EnvType, Environment}
-import net.fabricmc.fabric.api.client.render.BlockEntityRendererRegistry
-import notjoe.stockpile.blockentity.StockpileBarrelBlockEntity
-import notjoe.stockpile.renderer.StockpileBarrelRenderer
-
-@Environment(EnvType.CLIENT)
-class StockpileClientInitializer extends ClientModInitializer {
-  override def onInitializeClient(): Unit = {
-    BlockEntityRendererRegistry.INSTANCE
-      .register(classOf[StockpileBarrelBlockEntity], StockpileBarrelRenderer)
-  }
-}
