--- src/main/scala/notjoe/stockpile/block/StockpileProperties.scala
@@ -1,7 +1,0 @@
-package notjoe.stockpile.block
-
-import net.minecraft.state.property.BooleanProperty
-
-object StockpileProperties {
-  final val IsOpen = BooleanProperty.create("is_open")
-}
